s3:lib fix output of hex escapes > 0x7F with cbuf_print_quoted (cherry picked from commit a1fcb7e37a23dcc035686486d4afe32a67082ae1)

+++ source3/lib/cbuf.c
@@ -306,7 +306,7 @@ int cbuf_print_quoted(cbuf* ost, const char* s, size_t len)
 			if (isprint(*s) && ((*s == ' ') || !isspace(*s))) {
 				ret = cbuf_putc(ost, *s);
 			} else {
-				ret = cbuf_printf(ost, "\\%02x", *s);
+				ret = cbuf_printf(ost, "\\%02x", (uchar)*s);
 			}
 		}
 		s++;
